featL: add seperate player component
